+++ WhistleWind.AbnormalSigils/Abilities/abilityFrostRuler.cs
@@ -1,4 +1,5 @@
 ﻿using DiskCardGame;
+using InscryptionAPI.Card;
 using System.Collections;
 using UnityEngine;
 using WhistleWind.AbnormalSigils.Core.Helpers;
@@ -13,7 +14,7 @@ namespace WhistleWind.AbnormalSigils
         private void Ability_FrostRuler()
         {
             const string rulebookName = "Ruler of Frost";
-            const string rulebookDescription = "Once per turn, pay 3 Bones to choose a space on the board. If the space is occupied by a killable card, transform it into a Frozen Heart. Otherwise create a Block of Ice.";
+            const string rulebookDescription = "Once per turn, pay 2 Bones to choose a space on the board. If the space is empty, create a Block of Ice. Otherwise, if this card can kill the chosen card, transform it into a Frozen Heart.";
             const string dialogue = "With a wave of her hand, the Snow Queen blocked the path.";
             const string triggerText = "[creature] freezes the path.";
             FrostRuler.ability = AbnormalAbilityHelper.CreateActivatedAbility<FrostRuler>(
@@ -27,8 +28,17 @@ namespace WhistleWind.AbnormalSigils
         public static Ability ability;
         public override Ability Ability => ability;
         public override string NoTargetsDialogue => "The enemy is immune to the cold.";
-        public override string InvalidTargetDialogue => "Frost cannot penetrate this one. Choose another.";
-        public override int StartingBonesCost => 3;
+        public override string InvalidTargetDialogue(CardSlot slot)
+        {
+            if (slot.Card.Health > base.Card.Attack)
+                return "Its heart is too strong to ensare.";
+
+            if (slot.Card.HasAbility(Scorching.ability))
+                return "This creature is too firey to freeze.";
+
+            return "Frost cannot penetrate this one. Choose another.";
+        }
+        public override int StartingBonesCost => 2;
         public override int TurnDelay => 1;
 
         public override IEnumerator OnValidTargetSelected(CardSlot slot)
@@ -38,9 +48,7 @@ namespace WhistleWind.AbnormalSigils
             {
                 slot.Card.Anim.LightNegationEffect();
                 yield return new WaitForSeconds(0.15f);
-                yield return slot.Card.DieTriggerless();
-
-                yield return SpawnCard(slot, "wstl_snowQueenIceHeart");
+                yield return slot.Card.TransformIntoCard(CardLoader.GetCardByName("wstl_snowQueenIceHeart"), () => slot.Card.Status.damageTaken = 0);
                 yield return new WaitForSeconds(0.5f);
                 yield return DialogueHelper.PlayDialogueEvent("FrostRulerKiss");
             }
@@ -50,12 +58,15 @@ namespace WhistleWind.AbnormalSigils
             yield return base.LearnAbility();
         }
 
-        public override bool IsInvalidTarget(CardSlot slot)
+        public override bool IsValidTarget(CardSlot slot)
         {
             if (slot.Card == null)
+                return true;
+
+            if (slot.Card == base.Card || slot.Card.Dead)
                 return false;
 
-            return slot.Card.HasAbility(Scorching.ability) || slot.Card.Health > base.Card.Attack || slot.Card.Dead || slot.Card == base.Card;
+            return slot.Card.LacksAbility(Scorching.ability) && slot.Card.Health <= base.Card.Attack;
         }
 
         private IEnumerator SpawnCard(CardSlot slot, string name)