+++ WhistleWindLobotomyMod/Core/LobotomySaveManager.cs
@@ -1,4 +1,6 @@
-﻿using InscryptionAPI.Saves;
+﻿using DiskCardGame;
+using InscryptionAPI.Saves;
+using WhistleWindLobotomyMod.Challenges;
 
 namespace WhistleWindLobotomyMod.Core
 {
@@ -96,12 +98,8 @@ namespace WhistleWindLobotomyMod.Core
             get => GetSaveBool("UnlockedAngela");
             set => SetSaveValue("UnlockedAngela", value);
         }
-        public static bool DefeatedApocalypseBoss
-        {
-            get => GetSaveBool("DefeatedApocalypseBoss");
-            set => SetSaveValue("DefeatedApocalypseBoss", value);
-        }
-        public static bool DefeatedJesterBoss
+        public static bool DefeatedApocalypseBoss => AscensionSaveData.Data.conqueredChallenges.Contains(FinalApocalypse.Id);
+/*        public static bool DefeatedJesterBoss
         {
             get => GetSaveBool("DefeatedJesterBoss");
             set => SetSaveValue("DefeatedJesterBoss", value);
@@ -115,7 +113,7 @@ namespace WhistleWindLobotomyMod.Core
         {
             get => GetSaveBool("DefeatedSaviourBoss");
             set => SetSaveValue("DefeatedSaviourBoss", value);
-        }
+        }*/
         private static int GetRunInt(string id) => ModdedSaveManager.RunState.GetValueAsInt(LobotomyPlugin.pluginGuid, id);
         private static bool GetRunBool(string id) => ModdedSaveManager.RunState.GetValueAsBoolean(LobotomyPlugin.pluginGuid, id);
         private static void SetRunValue(string id, object value) => ModdedSaveManager.RunState.SetValue(LobotomyPlugin.pluginGuid, id, value);