+++ WhistleWindLobotomyMod/Opponents/Apocalypse/ApocalypseBossOpponent.cs
@@ -91,7 +91,6 @@ namespace WhistleWindLobotomyMod.Opponents.Apocalypse
 
             if (base.NumLives == 0)
             {
-                LobotomySaveManager.DefeatedApocalypseBoss = true;
                 AscensionStatsData.TryIncrementStat(AscensionStat.Type.BossesDefeated);
                 yield return new WaitForSeconds(0.25f);
                 Singleton<ViewManager>.Instance.SwitchToView(View.Default, immediate: false, lockAfter: true);